+++ tests/test_scheme3.py
@@ -1,4 +1,5 @@
 import time
+import numpy as np
 import spikeinterface as si
 import spikeinterface.extractors as se
 import spikeinterface.preprocessing as spre
@@ -21,7 +22,7 @@ def test_scheme3():
     timer = time.time()
 
     # lazy preprocessing
-    recording_filtered = spre.bandpass_filter(recording, freq_min=300, freq_max=6000, dtype=float)
+    recording_filtered = spre.bandpass_filter(recording, freq_min=300, freq_max=6000, dtype=np.float32)
     recording_preprocessed: si.BaseRecording = spre.whiten(recording_filtered)
 
     with TemporaryDirectory() as tmpdir:
